# Conflicts:
#	out/production/classes/com/example/project_machimo/auction/controller/AuctionRestController.class
#	out/production/classes/com/example/project_machimo/auction/dto/CheckDTO.class
#	out/production/resources/templates/fragments/header.html
#	src/main/resources/templates/fragments/header.html

+++ src/main/java/com/example/project_machimo/auction/dao/AuctionDAO.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Mapper
 public interface AuctionDAO {
     public AuctionDTO pList(int id);
-    public void highestBidUpdate(int amount,int id);
+    public void highestBidUpdate(Long amount,int id);
 
     public List<AuctionDTO> endList(Timestamp timestamp);
 }
